relocate helper file, update breakdown labels
